trashed stuff and begun rewriting of script
